combine all physics related system into one system and move collision detection system to ts
