--- src/contrib/systems/collision-detection-system/aabbBuilders/aabbBuilder.js
@@ -1,7 +1,0 @@
-class AABBBuilder {
-  getAABB() {
-    throw new Error('You should override this function');
-  }
-}
-
-export default AABBBuilder;
