--- src/contrib/systems/collision-detection-system/aabbBuilders/index.js
@@ -1,7 +1,0 @@
-import BoxAABBBuilder from './boxAABBBuilder';
-import CircleAABBBuilder from './circleAABBBuilder';
-
-export default {
-  boxCollider: BoxAABBBuilder,
-  circleCollider: CircleAABBBuilder,
-};
