--- src/contrib/systems/collision-detection-system/coordinatesCalculators/coordinatesCalculator.js
@@ -1,7 +1,0 @@
-class CoordinatesCalculator {
-  calc() {
-    throw new Error('You should override this function');
-  }
-}
-
-export default CoordinatesCalculator;
