--- src/contrib/systems/collision-detection-system/coordinatesCalculators/index.js
@@ -1,7 +1,0 @@
-import BoxCoordinatesCalculator from './boxCoordinatesCalculator';
-import CircleCoordinatesCalculator from './circleCoordinatesCalculator';
-
-export default {
-  boxCollider: BoxCoordinatesCalculator,
-  circleCollider: CircleCoordinatesCalculator,
-};
