--- src/contrib/systems/collision-detection-system/intersectionCheckers/index.js
@@ -1,10 +1,0 @@
-import BoxesIntersectionChecker from './boxesIntersectionChecker';
-import CirclesIntersectionChecker from './circlesIntersectionChecker';
-import BoxAndCircleIntersectionChecker from './boxAndCircleIntersectionChecker';
-
-export default {
-  boxCollider_boxCollider: BoxesIntersectionChecker,
-  circleCollider_circleCollider: CirclesIntersectionChecker,
-  circleCollider_boxCollider: BoxAndCircleIntersectionChecker,
-  boxCollider_circleCollider: BoxAndCircleIntersectionChecker,
-};
