--- src/contrib/systems/collision-detection-system/intersectionCheckers/intersectionChecker.js
@@ -1,7 +1,0 @@
-class IntersectionChecker {
-  check() {
-    throw new Error('You should override this function');
-  }
-}
-
-export default IntersectionChecker;
