--- src/contrib/systems/jammer/index.ts
@@ -1,27 +1,0 @@
-import type { System, SystemOptions } from '../../../engine/system';
-import type { MessageBus } from '../../../engine/message-bus';
-
-interface JammerOptions extends SystemOptions {
-  messages: Array<string>;
-}
-
-export class Jammer implements System {
-  private messageBus: MessageBus;
-  private messages: Array<string>;
-
-  constructor(options: SystemOptions) {
-    const {
-      messages,
-      messageBus,
-    } = options as JammerOptions;
-
-    this.messages = messages;
-    this.messageBus = messageBus;
-  }
-
-  update(): void {
-    this.messages.forEach((message) => {
-      this.messageBus.delete(message);
-    });
-  }
-}
