+++ src/contrib/systems/physics-system/consts.ts
@@ -1,0 +1,13 @@
+export const ADD_FORCE_MSG = 'ADD_FORCE';
+export const ADD_IMPULSE_MSG = 'ADD_IMPULSE';
+export const STOP_MOVEMENT_MSG = 'STOP_MOVEMENT';
+export const COLLISION_MSG = 'COLLISION';
+export const COLLISION_ENTER_MSG = 'COLLISION_ENTER';
+export const COLLISION_STAY_MSG = 'COLLISION_STAY';
+export const COLLISION_LEAVE_MSG = 'COLLISION_LEAVE';
+
+export const RIGID_BODY_COMPONENT_NAME = 'rigidBody';
+export const COLLIDER_COMPONENT_NAME = 'colliderContainer';
+export const TRANSFORM_COMPONENT_NAME = 'transform';
+
+export const GRAVITATIONAL_ACCELERATION_STORE_KEY = 'gravitationalAcceleration';
