+++ src/contrib/systems/physics-system/subsystems/collision-detection/aabb-builders/aabb-builder.ts
@@ -1,0 +1,8 @@
+import type { ColliderContainer } from '../../../../../components';
+import type { Coordinates } from '../types';
+
+import type { AABB } from './aabb';
+
+export interface AABBBuilder {
+  getAABB(collider: ColliderContainer, coordinates: Coordinates): AABB
+}
