+++ src/contrib/systems/physics-system/subsystems/collision-detection/aabb-builders/aabb.ts
@@ -1,0 +1,11 @@
+import type { Point } from '../types';
+
+export class AABB {
+  min: Point;
+  max: Point;
+
+  constructor(min: Point, max: Point) {
+    this.min = min;
+    this.max = max;
+  }
+}
