+++ src/contrib/systems/physics-system/subsystems/collision-detection/aabb-builders/index.ts
@@ -1,0 +1,11 @@
+import { BoxAABBBuilder } from './box-aabb-builder';
+import { CircleAABBBuilder } from './circle-aabb-builder';
+import type { AABBBuilder } from './aabb-builder';
+
+export type { AABBBuilder };
+export type { AABB } from './aabb';
+
+export const aabbBuilders: Record<string, { new(): AABBBuilder }> = {
+  boxCollider: BoxAABBBuilder,
+  circleCollider: CircleAABBBuilder,
+};
