+++ src/contrib/systems/physics-system/subsystems/collision-detection/coordinates-calculators/coordinates-calculator.ts
@@ -1,0 +1,9 @@
+import type {
+  ColliderContainer,
+  Transform,
+} from '../../../../../components';
+import type { Coordinates } from '../types';
+
+export interface CoordinatesCalculator {
+  calc(collider: ColliderContainer, transform: Transform): Coordinates
+}
