+++ src/contrib/systems/physics-system/subsystems/collision-detection/intersection-checkers/intersection-checker.ts
@@ -1,0 +1,16 @@
+import type { Vector2 } from '../../../../../../engine/mathLib';
+import type { ColliderContainer } from '../../../../../components';
+import type { Coordinates } from '../types';
+
+export interface Intersection {
+  mtv1: Vector2
+  mtv2: Vector2
+}
+
+export interface IntersectionEntry extends Pick<ColliderContainer, 'type' | 'collider'> {
+  coordinates: Coordinates
+}
+
+export interface IntersectionChecker {
+  check(arg1: IntersectionEntry, arg2: IntersectionEntry): Intersection | false
+}
