+++ src/contrib/systems/physics-system/subsystems/index.ts
@@ -1,0 +1,5 @@
+export { PhysicsSubsystem } from './physics';
+export { CollisionDetectionSubsystem } from './collision-detection';
+export { CollisionBroadcastSubsystem } from './collision-broadcast';
+export { CollisionSolver } from './collision-solver';
+export { ConstraintSolver } from './constraint-solver';
